added functions for step 5&6

+++ src/edu/upenn/cit594/processor/AverageCalculator.java
@@ -4,5 +4,5 @@ import java.util.ArrayList;
 import edu.upenn.cit594.data.Property;
 
 public interface AverageCalculator {
-	public Double calculateAverage (ArrayList<Property> properties) ;
+	public int calculateAverage (ArrayList<Property> properties) ;
 }
